worker: move CsmockRunner class to the worker module
Resolves: https://gitlab.cee.redhat.com/covscan/covscan/-/issues/268
