+++ osh/worker/tasks/task_errata_diff_build.py
@@ -6,7 +6,7 @@ import sys
 
 from kobo.worker import TaskBase
 
-from osh.common.csmock_parser import CsmockRunner
+from osh.worker.csmock_runner import CsmockRunner
 
 
 class ErrataDiffBuild(TaskBase):
